# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "Platform Import Export API",
                 "version":  "1.0"
             },
    "paths":  {
...
                  "/api/v1/messaging/conversations":  {
                                                          "get":  {
                                                                      "tags":  [
                                                                                   "messaging",
                                                                                   "Messaging"
                                                                               ],
                                                                      "summary":  "List Conversations",
                                                                      "operationId":  "list_conversations_api_v1_messaging_conversations_get",
                                                                      "responses":  {
                                                                                        "200":  {
                                                                                                    "description":  "Successful Response",
                                                                                                    "content":  {
                                                                                                                    "application/json":  {
                                                                                                                                             "schema":  {
                                                                                                                                                            "items":  {
                                                                                                                                                                          "$ref":  "#/components/schemas/ConversationOut"
                                                                                                                                                                      },
                                                                                                                                                            "type":  "array",
                                                                                                                                                            "title":  "Response List Conversations Api V1 Messaging Conversations Get"
                                                                                                                                                        }
                                                                                                                                         }
                                                                                                                }
                                                                                                }
                                                                                    },
                                                                      "security":  [
                                                                                       {
                                                                                           "HTTPBearer":  [

                                                                                                          ]
                                                                                       }
                                                                                   ]
                                                                  },
                                                          "post":  {
                                                                       "tags":  [
                                                                                    "messaging",
                                                                                    "Messaging"
                                                                                ],
                                                                       "summary":  "Initiate Conversation",
                                                                       "operationId":  "initiate_conversation_api_v1_messaging_conversations_post",
                                                                       "requestBody":  {
                                                                                           "content":  {
                                                                                                           "application/json":  {
                                                                                                                                    "schema":  {
                                                                                                                                                   "$ref":  "#/components/schemas/ConversationCreate"
                                                                                                                                               }
                                                                                                                                }
                                                                                                       },
                                                                                           "required":  true
                                                                                       },
                                                                       "responses":  {
                                                                                         "200":  {
                                                                                                     "description":  "Successful Response",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/ConversationOut"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 },
                                                                                         "422":  {
                                                                                                     "description":  "Validation Error",
                                                                                                     "content":  {
                                                                                                                     "application/json":  {
                                                                                                                                              "schema":  {
                                                                                                                                                             "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                         }
                                                                                                                                          }
                                                                                                                 }
                                                                                                 }
                                                                                     },
                                                                       "security":  [
                                                                                        {
                                                                                            "HTTPBearer":  [

                                                                                                           ]
                                                                                        }
                                                                                    ]
                                                                   }
                                                      },
                  "/api/v1/messaging/conversations/{conversation_id}/messages":  {
                                                                                     "get":  {
                                                                                                 "tags":  [
                                                                                                              "messaging",
                                                                                                              "Messaging"
                                                                                                          ],
                                                                                                 "summary":  "Get Messages",
                                                                                                 "operationId":  "get_messages_api_v1_messaging_conversations__conversation_id__messages_get",
                                                                                                 "security":  [
                                                                                                                  {
                                                                                                                      "HTTPBearer":  [

                                                                                                                                     ]
                                                                                                                  }
                                                                                                              ],
                                                                                                 "parameters":  [
                                                                                                                    {
                                                                                                                        "name":  "conversation_id",
                                                                                                                        "in":  "path",
                                                                                                                        "required":  true,
                                                                                                                        "schema":  {
                                                                                                                                       "type":  "string",
                                                                                                                                       "title":  "Conversation Id"
                                                                                                                                   }
                                                                                                                    }
                                                                                                                ],
                                                                                                 "responses":  {
                                                                                                                   "200":  {
                                                                                                                               "description":  "Successful Response",
                                                                                                                               "content":  {
                                                                                                                                               "application/json":  {
                                                                                                                                                                        "schema":  {
                                                                                                                                                                                       "type":  "array",
                                                                                                                                                                                       "items":  {
                                                                                                                                                                                                     "$ref":  "#/components/schemas/MessageOut"
                                                                                                                                                                                                 },
                                                                                                                                                                                       "title":  "Response Get Messages Api V1 Messaging Conversations  Conversation Id  Messages Get"
                                                                                                                                                                                   }
                                                                                                                                                                    }
                                                                                                                                           }
                                                                                                                           },
                                                                                                                   "422":  {
                                                                                                                               "description":  "Validation Error",
                                                                                                                               "content":  {
                                                                                                                                               "application/json":  {
                                                                                                                                                                        "schema":  {
                                                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                                   }
                                                                                                                                                                    }
                                                                                                                                           }
                                                                                                                           }
                                                                                                               }
                                                                                             }
                                                                                 },
                  "/api/v1/messaging/conversations/{conversation_id}/status":  {
                                                                                   "put":  {
                                                                                               "tags":  [
                                                                                                            "messaging",
                                                                                                            "Messaging"
                                                                                                        ],
                                                                                               "summary":  "Update Status",
                                                                                               "operationId":  "update_status_api_v1_messaging_conversations__conversation_id__status_put",
                                                                                               "security":  [
                                                                                                                {
                                                                                                                    "HTTPBearer":  [

                                                                                                                                   ]
                                                                                                                }
                                                                                                            ],
                                                                                               "parameters":  [
                                                                                                                  {
                                                                                                                      "name":  "conversation_id",
                                                                                                                      "in":  "path",
                                                                                                                      "required":  true,
                                                                                                                      "schema":  {
                                                                                                                                     "type":  "string",
                                                                                                                                     "title":  "Conversation Id"
                                                                                                                                 }
                                                                                                                  }
                                                                                                              ],
                                                                                               "requestBody":  {
                                                                                                                   "required":  true,
                                                                                                                   "content":  {
                                                                                                                                   "application/json":  {
                                                                                                                                                            "schema":  {
                                                                                                                                                                           "$ref":  "#/components/schemas/ConversationUpdateStatus"
                                                                                                                                                                       }
                                                                                                                                                        }
                                                                                                                               }
                                                                                                               },
                                                                                               "responses":  {
                                                                                                                 "200":  {
                                                                                                                             "description":  "Successful Response",
                                                                                                                             "content":  {
                                                                                                                                             "application/json":  {
                                                                                                                                                                      "schema":  {
                                                                                                                                                                                     "$ref":  "#/components/schemas/ConversationOut"
                                                                                                                                                                                 }
                                                                                                                                                                  }
                                                                                                                                         }
                                                                                                                         },
                                                                                                                 "422":  {
                                                                                                                             "description":  "Validation Error",
                                                                                                                             "content":  {
                                                                                                                                             "application/json":  {
                                                                                                                                                                      "schema":  {
                                                                                                                                                                                     "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                                                 }
                                                                                                                                                                  }
                                                                                                                                         }
                                                                                                                         }
                                                                                                             }
                                                                                           }
                                                                               },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "ConversationOut":  {
                                                               "properties":  {
                                                                                  "listing_id":  {
                                                                                                     "type":  "string",
                                                                                                     "title":  "Listing Id"
                                                                                                 },
                                                                                  "id":  {
                                                                                             "type":  "string",
                                                                                             "title":  "Id"
                                                                                         },
                                                                                  "initiator_company_id":  {
                                                                                                               "type":  "string",
                                                                                                               "title":  "Initiator Company Id"
                                                                                                           },
                                                                                  "statut":  {
                                                                                                 "$ref":  "#/components/schemas/StatutConversation"
                                                                                             },
                                                                                  "date_creation":  {
                                                                                                        "type":  "string",
                                                                                                        "format":  "date-time",
                                                                                                        "title":  "Date Creation"
                                                                                                    },
                                                                                  "date_maj":  {
                                                                                                   "anyOf":  [
                                                                                                                 {
                                                                                                                     "type":  "string",
                                                                                                                     "format":  "date-time"
                                                                                                                 },
                                                                                                                 {
                                                                                                                     "type":  "null"
                                                                                                                 }
                                                                                                             ],
                                                                                                   "title":  "Date Maj"
                                                                                               }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "listing_id",
                                                                                "id",
                                                                                "initiator_company_id",
                                                                                "statut",
                                                                                "date_creation",
                                                                                "date_maj"
                                                                            ],
                                                               "title":  "ConversationOut"
                                                           },
...
                                   },
                       "securitySchemes":  {
                                               "HTTPBearer":  {
                                                                  "type":  "http",
                                                                  "scheme":  "bearer"
                                                              }
                                           }
                   }
}